@actionless 2015-09-20
now it's fine -- no eye-observable performance issues and problem with fitting new size when child items changed not reproducible no more