@blueyed 2015-09-20
For some reason the Travis build does not start, although I've stopped and restarted it (https://travis-ci.org/awesomeWM/awesome/builds/81255650).
Trying to close and re-open the PR to trigger it.