@psychon 2015-09-20
@actionless Could you tell me more about the previous performance issues? I'd like to look into them.
(Also, awesome seems to have worse memory usage now, but debugging claims that memory usage does not increase, so no leaks. Anyone else sees something?)
Finally: @blueyed https://travis-ci.org/awesomeWM/awesome/pull_requests #452 seems to have travis blocked.