@actionless 2015-09-20
i have a pretty complicated layout of widgets composed from flex, fixed, constraint etc layouts on my panel and each tag switch was happened slowly because of panel redrawing